separate errors to their own type
