+++ packages/two-factor/src/types/error-messages.ts
@@ -1,0 +1,7 @@
+export interface ErrorMessages {
+  userNotFound: string;
+  codeDidNotMatch: string;
+  userTwoFactorNotSet: string;
+  userTwoFactorAlreadySet: string;
+  codeRequired: string;
+}
