+++ packages/two-factor/src/types/options.ts
@@ -1,10 +1,4 @@
-export interface ErrorMessages {
-  userNotFound: string;
-  codeDidNotMatch: string;
-  userTwoFactorNotSet: string;
-  userTwoFactorAlreadySet: string;
-  codeRequired: string;
-}
+import { ErrorMessages } from './error-messages';
 
 export interface AccountsTwoFactorOptions {
   /**
